backend/requirements.txt
google-generativeai
fastapi
uvicorn
python-dotenv
pydantic
gunicorn
